(v0.3.2) Features added
(29.08.2019)
 - Added chicken multibuy
 - Some fixes and improvements

+++ chickenfight/Assets/Scripts/GlobalCash.cs
@@ -14,7 +14,7 @@ public class GlobalCash : MonoBehaviour
     void Update()
     {
         InternalCash = CashCount;
-        totalCash.GetComponent<Text>().text = "AMOUNT OF CASH: " + (int)InternalCash;
+        totalCash.GetComponent<Text>().text = "AMOUNT OF CASH: " + InternalCash;
     }
